Refactor dashboard into package
